+++ common/src/main/resources/assets/mcui/themes/hex2/scripts/widgetlib.lua
@@ -142,7 +142,7 @@ function wl.loadButton(parent, args)
     wl.bind(baseArgs, 'text', args.label, wl.tstatic, 'STRING')
     wl.bind(baseArgs, 'xPos', args.xPos, wl.tstatic, 'DOUBLE')
     wl.bind(baseArgs, 'yPos', args.yPos, wl.tstatic, 'DOUBLE')
-    wl.bind(baseArgs, 'initialWidth', args.width or 160, wl.tstatic, 'DOUBLE')
+    wl.bind(baseArgs, 'initialWidth', args.width or 160, wl.tstatic, 'INT')
 
     local allArgs = util.merge(
             baseArgs,
